refactor(web): delete use-ai-simple-mode and use-registry-settings hooks
Migrates the remaining three registry consumers
(use-install-from-registry, use-enabled-registries, use-registry-connections)
to the unified useOrganizationSettings hook and its useIsRegistryEnabled /
useRegistryConfig wrappers, then deletes the two legacy hook files.
